Refactor timer class
